add api functions for weather/location data
this commit sets up rough inputs on a webpage for a user to search
any location and recieve weather information for that location logged
to the console. additinally on the initial load the page will try to
find the user location and return weather information for their
location. on nay subsequent reloads of the page the last searched
location will load in, although this feature is not finished